More Route additions and transition migration to the Node schema
Added editing of stories and nodes within stories. Currently there is a bug where the console does not display the correct updated data for node editing, but it appears the data is still inserted into the database correctly so it has been ignored for that. Also, migrated the Story schema's transition table into a transition array for the Node schema.