great progress towards tree visualization
CSS is relatively understood and in order to get working prototype going. On the HTML, I have 3 different trees build to teest and play with how to make them. the final tree is where I slowly build how and edited the CSS for a nice looking flow and proper box sizes.